feat: modified dummy_data command to create category and expertise
